Bug 469523 - Enable TUnit leak log in tinderbox (log); (Ev1) Support xpcshell built without leak logging; (Thunderbird bustage fix)
UltraBlame original commit: 46a0b3d6c672c11dd7a1103da90be7b4b78bba86

+++ testing/xpcshell/runxpcshelltests.py
@@ -194,7 +194,9 @@ def runTests(xpcshell, testdirs=[], xrePath=None, testFile=None,
         print "TEST-PASS | %s | all tests passed" % test
       processLeakLog(leakLogFile)
       
-      os.remove(leakLogFile)
+      
+      if os.path.exists(leakLogFile):
+        os.remove(leakLogFile)
       if not (success or keepGoing):
         return False
